[Rostest] remove the wrong duplicated file for hovering check
